Move retry logic to decorator defined in utils.py
